Fix SUSE E2E timeout failures.
The import of SUSE images has started taking longer than the default
timeout limit of 2h. To make our tests more stable we can increase this
value temporary.

In the future we can work on the import performance, for example, when
we will migrate the import / export tools on N2 instances, or in the
context of a different ticket.